Refactor: Remove MyPage, add Bookmark.jsx, and relocate bookmark components to bookmark folder
